working on resources and getting testing working
