+++ .github/workflows/test.yml
@@ -55,9 +55,7 @@ jobs:
       matrix:
         # list whatever Terraform versions here you would like to support
         terraform:
-          - '1.0.*'
-          - '1.1.*'
-          - '1.2.*'
+          - '1.3.*'
     steps:
       - uses: actions/checkout@v3
       - uses: actions/setup-go@v3
