--- internal/jumpcloud/kv/kv_model.go
@@ -1,11 +1,0 @@
-package kv
-
-import (
-	"github.com/hashicorp/terraform-plugin-framework/tfsdk"
-	"github.com/hashicorp/terraform-plugin-framework/types"
-)
-
-type KVItemModel struct {
-	Name					types.String						`tfsdk:"name"`
-	Value					types.String						`tfsdk:"value"`
-}
