+++ internal/jumpcloud/kv_model.go
@@ -1,0 +1,10 @@
+package jumpcloud
+
+import (
+	"github.com/hashicorp/terraform-plugin-framework/types"
+)
+
+type KVItemModel struct {
+	Name  types.String `tfsdk:"name"`
+	Value types.String `tfsdk:"value"`
+}
