+++ internal/jumpcloud/provider_test.go
@@ -1,0 +1,18 @@
+package jumpcloud
+
+import (
+	"github.com/hashicorp/terraform-plugin-framework/providerserver"
+	"github.com/hashicorp/terraform-plugin-go/tfprotov6"
+)
+
+const ProviderConfig = `
+variable "jumpcloud_api_key" {}
+
+provider "jumpcloud" {
+	api_key = "${ var.jumpcloud_api_key }"
+}
+`
+
+var testAccProtoV6ProviderFactories = map[string]func() (tfprotov6.ProviderServer, error){
+	"jumpcloud": providerserver.NewProtocol6WithError(New("dev")()),
+}
