--- internal/jumpcloud/usergroup/usergroup_datasource.go
@@ -1,6 +1,0 @@
-package usergroup
-
-import (
-
-)
-
